Document the new 'feature'
