+++ examples/basic/main.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"errors"
+	"log/slog"
+	"os"
+
+	"github.com/phsym/console-slog"
+)
+
+func main() {
+	logger := slog.New(
+		console.NewHandler(os.Stderr, &console.HandlerOptions{Level: slog.LevelDebug, AddSource: true}),
+	)
+	slog.SetDefault(logger)
+	slog.Info("Hello world!", "foo", "bar")
+	slog.Debug("Debug message")
+	slog.Warn("Warning message")
+	slog.Error("Error message", "err", errors.New("the error"))
+
+	logger = logger.With("foo", "bar").
+		WithGroup("the-group").
+		With("bar", "baz")
+
+	logger.Info("group info", "attr", "value")
+}
